refactor(증정 기능) : GiftResults 생성 방식 변경
LocalDate 객체와 SelectedMenus 객체를 이용해 DiscountResults를 생성하는 메서드는 StatisticsGenerator 내부에 있는 것 보다 DiscountResults 내부에 있는 것이 더 적절하다. 이 메서드를 정적 메서드로 추가하여 SelectedMenus와 독립적으로 동작할 수 있도록 하였다.
