+++ src/main/java/christmas/domain/discount/DiscountResults.java
@@ -1,5 +1,8 @@
 package christmas.domain.discount;
 
+import christmas.domain.menu.SelectedMenus;
+import java.time.LocalDate;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -8,6 +11,22 @@ public final class DiscountResults {
 
     public DiscountResults(List<DiscountResult> discountResults) {
         this.discountResults = discountResults;
+    }
+
+    public static DiscountResults calculatedFrom(LocalDate selectedDate, SelectedMenus selectedMenus) {
+        List<DiscountResult> discountResults = calculateDiscountResults(selectedDate, selectedMenus);
+        return new DiscountResults(discountResults);
+    }
+
+    private static List<DiscountResult> calculateDiscountResults(LocalDate selectedDate, SelectedMenus selectedMenus) {
+        return Arrays.stream(DiscountType.values())
+                .map(discountType -> calculateDiscountResult(selectedDate, selectedMenus, discountType)).toList();
+    }
+
+    private static DiscountResult calculateDiscountResult(LocalDate selectedDate, SelectedMenus selectedMenus,
+                                                          DiscountType discountType) {
+        DiscountCalculator discountCalculator = DiscountCalculatorMapper.map(discountType).orElseThrow();
+        return discountCalculator.calculate(selectedDate, selectedMenus);
     }
 
     public int getTotalDiscountPrice() {
